Separate bash scripts into own file and use classmethods where appropriate
